requirements.txt
accelerate==1.2.1
BOND==1.1.0
datasets==2.21.0
numpy==2.2.6
rich==14.0.0
stefutils==0.43.0
torch==2.3.1
tqdm==4.66.4
transformers==4.51.3
wandb==0.19.11
